+++ 031602408/wordCount/src/input.txt
@@ -1,60 +1,10 @@
-that you will be more mild and tractable .
-and fear not , lords , and you , lavinia ,
-by my advice , all humbled on your knees ,
-you shall ask pardon of his majesty .
-
-we do ; and vow to heaven and to his highness ,
-that what we did was mildly , as we might ,
-tendering our sister's honour and our own .
-
-that on mine honour here i do protest .
-
-away , and talk not ; trouble us no more .
-
-nay , nay , sweet emperor , we must all be friends :
-the tribune and his nephews kneel for grace ;
-i will not be denied : sweet heart , look back .
-
-marcus , for thy sake , and thy brother's here ,
-and at my lovely tamora's entreats ,
-i do remit these young men's heinous faults :
-stand up .
-lavinia , though you left me like a churl ,
-i found a friend , and sure as death i swore 
-i would not part a bachelor from the priest .
-come ; if the emperor's court can feast two brides ,
-you are my guest , lavinia , and your friends .
-this day shall be a love-day , tamora .
-
-to-morrow , an it please your majesty 
-to hunt the panther and the hart with me ,
-with horn and hound we'll give your grace bon jour .
-
-be it so , titus , and gramercy too .
-
-
-now climbeth tamora olympus' top ,
-safe out of fortune's shot ; and sits aloft ,
-secure of thunder's crack or lightning flash ,
-advanc'd above pale envy's threat'ning reach .
-as when the golden sun salutes the morn ,
-and , having gilt the ocean with his beams ,
-gallops the zodiac in his glistering coach ,
-and overlooks the highest-peering hills ;
-so tamora .
-upon her wit doth earthly honour wait 
-and virtue stoops and trembles at her frown .
-then , aaron , arm thy heart , and fit thy thoughts 
-to mount aloft with thy imperial mistress ,
-and mount her pitch , whom thou in triumph long 
-hast prisoner held , fetter'd in amorous chains ,
-and faster bound to aaron's charming eyes 
-than is prometheus tied to caucasus .
-away with slavish weeds and servile thoughts !
-i will be bright , and shine in pearl and gold ,
-to wait upon this new-made empress .
-to wait , said i ? to wanton with this queen ,
-this goddess , this semiramis , this nymph ,
-this siren , that will charm rome's saturnine ,
-and see his shipwrack and his commonweal's .
-holla ! what storm is this ?
+a
+a
+a
+abcd
+abcd1234
+abcd
+ abcd
+file
+file
+123file
