+++ 031602408/wordCount/src/result.txt
@@ -1,0 +1,6 @@
+characters: 49
+words: 7
+lines: 10
+<abcd>: 3
+<file>: 3
+<abcd1234>: 1
