--- 031602408/wordCount/src/wordCount.txt
@@ -1,10 +1,0 @@
-a
-a
-a
-abcd
-abcd1234
-abcd
- abcd
-file
-file
-123file
